Refactoring sales
